scrape: add options to manager to allow scraping at startup and shutdown
This change adds the following:
- `ScrapeOnShutdown` option - when enabled will force a final scrape
  before tearing down
- `InitialScrapeOffset` option - when set, all targets will be scraped
  at the defined offset

These options are useful for serverless flavours of the OTel collector
that uses the scrape manager. In serverless environments, instances are
short lived and we deploy the collector as a sidecar. Making sure the
scraping is not sensitive to startup latencies is important - as is a
guarantee that there will be atleast one scrape before the instance is
killed.

Signed-off-by: Ridwan Sharif <ridwanmsharif@google.com>